Added simple API generator and registration. Stable build, cleanup

+++ wrapper/src/main/kotlin/godot/core/CoreType.kt
@@ -10,10 +10,10 @@ internal interface CoreType {
 
 
 fun Int.getRawMemory(memScope: MemScope): COpaquePointer {
-    return cValuesOf(this.toLong()).getPointer(memScope)
+    return cValuesOf(this).getPointer(memScope)
 }
 fun Float.getRawMemory(memScope: MemScope): COpaquePointer {
-    return cValuesOf(this.toDouble()).getPointer(memScope)
+    return cValuesOf(this).getPointer(memScope)
 }
 fun Boolean.getRawMemory(memScope: MemScope): COpaquePointer {
     return cValuesOf(this.toByte()).getPointer(memScope)
